+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add a `-first` modifier class to the first visible entry of a table-of-contents element
 * Make HTML widget editable
+* Also store and retrieve copied widget config from localStorage
 
 ## 0.2.0 (2022-11-02)